+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Models/Exceptions.cs
@@ -1,3 +1,4 @@
+using KRT.VRCQuestTools.Components;
 using UnityEngine;
 
 #pragma warning disable SA1402 // file may only contain a single type
@@ -6,10 +7,17 @@ using UnityEngine;
 namespace KRT.VRCQuestTools.Models
 {
     /// <summary>
+    /// Interface for VRCQuestTools exceptions.
+    /// </summary>
+    internal interface IVRCQuestToolsException
+    {
+    }
+
+    /// <summary>
     /// Exception for conversion error.
     /// </summary>
     /// <typeparam name="T">Source object type.</typeparam>
-    internal abstract class ConversionException<T> : System.Exception
+    internal abstract class ConversionException<T> : System.Exception, IVRCQuestToolsException
     {
         /// <summary>
         /// Source object.
@@ -81,9 +89,39 @@ namespace KRT.VRCQuestTools.Models
     }
 
     /// <summary>
+    /// Exception for invalid material swap which contains null references.
+    /// </summary>
+    internal class InvalidMaterialSwapNullException : System.Exception, IVRCQuestToolsException
+    {
+        /// <summary>
+        /// Component which has the error.
+        /// </summary>
+        public readonly MaterialSwap component;
+
+        private readonly int index;
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="InvalidMaterialSwapNullException"/> class.
+        /// </summary>
+        /// <param name="message">Message.</param>
+        /// <param name="materialSwap">Component which has the error.</param>
+        /// <param name="index">Index of the material mapping which has the error.</param>
+        public InvalidMaterialSwapNullException(string message, MaterialSwap materialSwap, int index)
+            : base(message)
+        {
+            this.component = materialSwap;
+        }
+
+        /// <summary>
+        /// Gets the material mapping which has the error.
+        /// </summary>
+        public MaterialSwap.MaterialMapping MaterialMapping => component.materialMappings[index];
+    }
+
+    /// <summary>
     /// Exception for replacement material is not allowed for mobile.
     /// </summary>
-    internal class InvalidReplacementMaterialException : System.Exception
+    internal class InvalidReplacementMaterialException : System.Exception, IVRCQuestToolsException
     {
         /// <summary>
         /// Component which has the error.