+++ CHANGELOG.md
@@ -8,10 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Error case for missing materials in `VQT Material Swap` component.
 - [NDMF] Add `NDMF Phase` option to `VQT Mesh Flipper` component.
     - Before Decimation
     - After Decimation
 - [NDMF] NDMF Preview for `VQT Mesh Flipper`.
+
+### Fixed
+- Unhandled exception when replaced material is not set in `VQT Avatar Converter Settings`.
+- [NDMF] Animation conversion error not reported to the NDMF Console.
+- [NDMF] Material conversion error detail was wrongly reported in the NDMF Console.
 
 ## [2.8.2] - 2025-04-01