+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
+- Unhandled exception when replaced material is not set in `VQT Avatar Converter Settings`.
 - [NDMF] Animation conversion error not reported to the NDMF Console.
 - [NDMF] Material conversion error detail was wrongly reported in the NDMF Console.